feat: added primitive agent for managing Linux and extending MDM protocol
The agent is a binary that is put on the target machine and it helps to manage the machine. This commit builds a super basic agent. In the future authentication and a user interface will need to be added, along with a daemon to handle management syncing on a timer.