feat: refactor for multitenant support + linted
The server has been refactor to allow for multiple tenants to be hosted on the same server. The project is changing focus to this over single instance deployments for the Mattrax Cloud a hosted version of Mattrax. This will be a good way to test and fix any issues in the early rollout of the project. All files have also been linted and given a LF end of line sequence.