feat: upgraded applications resource + general improvements
The error page now includes the Zipkin request ID to aid with debugging. The settings overview page has been overhauled to include extra useful information about the state of the server and the resources it relies on. Applications were added to the API and connected to the frontend. A resource called objects was also added which is used for storing bulk files. This is currently used by the applications resource to store the MSI files, but may be used by other resources in the future.