feat: development instructions
The README now has instructions for how to run the code. Docker is used to run the Postgres database and the rest is run natively on the host to aid with debugging and development speed. When a blank DB is found a default admin user is created and the random password is printed to the console.