feat: refactored MDM protocols + added Android management support
The MDM protocol implementation has been overhauled to allow for Android For Work support. The enrollment UI has been updated to support the various platforms. The policy format has changed slightly to support the new implementation. Mattrax Policies are replicated to the Android Management API on creation, updates and deletion.